Fullstack Changes for using Postgres DD and JWT auth

+++ backend/TodoFullStack/src/main/java/com/swarna/todoFullStack/todo/TodoHardCodedService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 public class TodoHardCodedService {
 
 	private static List<Todo> todos = new ArrayList<>();
-	private static int idCounter = 0;
+	private static Long idCounter = (long) 0;
 
 	static {
 		todos.add(new Todo(++idCounter, "user", "My first Todo", new Date(), false));
